add dynamic import for cv2 and pyrealsense2
